gh-1798 us dl/avgdl in bm25 calculation
also remove doc frequency from regular where filter, where no scoring
occurs

+++ adapters/repos/db/inverted/analyzer_test.go
@@ -31,27 +31,27 @@ func TestAnalyzer(t *testing.T) {
 			assert.ElementsMatch(t, res, []Countable{
 				{
 					Data:          []byte("hello"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 				{
 					Data:          []byte("my"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 				{
 					Data:          []byte("name"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 				{
 					Data:          []byte("is"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 				{
 					Data:          []byte("john"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 				{
 					Data:          []byte("doe"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 			})
 		})
@@ -61,19 +61,19 @@ func TestAnalyzer(t *testing.T) {
 			assert.ElementsMatch(t, res, []Countable{
 				{
 					Data:          []byte("du"),
-					TermFrequency: float64(4),
+					TermFrequency: float32(4),
 				},
 				{
 					Data:          []byte("hast"),
-					TermFrequency: float64(3),
+					TermFrequency: float32(3),
 				},
 				{
 					Data:          []byte("mich"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 				{
 					Data:          []byte("gefragt"),
-					TermFrequency: float64(1),
+					TermFrequency: float32(1),
 				},
 			})
 		})
@@ -84,19 +84,19 @@ func TestAnalyzer(t *testing.T) {
 		assert.ElementsMatch(t, res, []Countable{
 			{
 				Data:          []byte("john-thats-jay.ohh.age.n+alloneword@doe.com"),
-				TermFrequency: float64(1),
+				TermFrequency: float32(1),
 			},
 			{
 				Data:          []byte("My"),
-				TermFrequency: float64(1),
+				TermFrequency: float32(1),
 			},
 			{
 				Data:          []byte("email"),
-				TermFrequency: float64(1),
+				TermFrequency: float32(1),
 			},
 			{
 				Data:          []byte("is"),
-				TermFrequency: float64(1),
+				TermFrequency: float32(1),
 			},
 		})
 	})
